+++ tests_and_analysis/test/euphonic_test/test_spectrum1d.py
@@ -85,7 +85,10 @@ def get_expected_spectrum1d(json_filename):
     return ExpectedSpectrum1D(get_spectrum1d_path(json_filename))
 
 
-def check_spectrum1d(actual_spectrum1d, expected_spectrum1d):
+def check_spectrum1d(actual_spectrum1d, expected_spectrum1d, y_atol=None):
+
+    if y_atol is None:
+        y_atol = np.finfo(np.float64).eps
 
     assert (actual_spectrum1d.x_data.units
             == expected_spectrum1d.x_data.units)
@@ -97,7 +100,7 @@ def check_spectrum1d(actual_spectrum1d, expected_spectrum1d):
             == expected_spectrum1d.y_data.units)
     npt.assert_allclose(actual_spectrum1d.y_data.magnitude,
                         expected_spectrum1d.y_data.magnitude,
-                        atol=np.finfo(np.float64).eps)
+                        atol=y_atol)
 
     if expected_spectrum1d.x_tick_labels is None:
         assert actual_spectrum1d.x_tick_labels is None