+++ tests_and_analysis/test/utils.py
@@ -95,6 +95,12 @@ def get_test_qpts(qpts_option: Optional[str] = 'default') -> np.ndarray:
             [2.00, 2.00, 0.5],
             [1.75, 0.50, 2.50]])
 
+def get_mode_widths(fc_path):
+    with open(os.path.join(fc_path), 'r') as fp:
+        modw_dict = json.load(fp)
+    mode_widths = modw_dict['mode_widths']*ureg(
+        modw_dict['mode_widths_unit'])
+    return mode_widths
 
 def check_frequencies_at_qpts(qpts, freqs, expected_freqs, atol, rtol,
                               gamma_atol=None):